Reduce code duplication across the codebase
- Extract GitHub issue URL parsing into lib/github.ts
- Split lib/utils.ts so pure functions live in lib/shared-utils.ts
  (client-safe), removing duplicated copies from ProjectList
- Extract shared input/button class constants into lib/styles.ts
