+++ lib/utils.ts
@@ -1,6 +1,9 @@
 import { auth } from "@/lib/auth";
 import { redirect } from "next/navigation";
 
+// Re-export pure utilities so existing server-side imports keep working.
+export { isProjectOpen, formatDateAsDaysInPast, isValidEmail } from "@/lib/shared-utils";
+
 /**
  * Call this at the top of any page that requires the user to have
  * completed setup (i.e. provided a contact email). If they're logged
@@ -12,28 +15,4 @@ export async function requireSetup() {
         redirect("/setup");
     }
     return session;
-}
-
-export function isValidEmail(email: string): boolean {
-    return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
-}
-
-/** A project is "open" when it has no student, isn't completed, and the mentor is available. */
-export function isProjectOpen(project: { studentId: bigint | null; completedAt: Date | null; mentorAvailable: boolean }): boolean {
-    return !project.studentId && !project.completedAt && project.mentorAvailable;
-}
-
-export function formatDateAsDaysInPast(date: Date) {
-    const now = new Date();
-    const lastSeenDay = new Date(date.getFullYear(), date.getMonth(), date.getDate());
-    const nowDay = new Date(now.getFullYear(), now.getMonth(), now.getDate());
-
-    if (lastSeenDay > nowDay) {
-        return "in the future (timey wimey stuff going on here...)";
-    } else if (lastSeenDay.getTime() === nowDay.getTime()) {
-        return "today";
-    } else {
-        const daysAgo = Math.round((nowDay.getTime() - lastSeenDay.getTime()) / (1000 * 60 * 60 * 24));
-        return `${daysAgo} day${daysAgo > 1 ? "s" : ""} ago`;
-    }
 }
